requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
coverage==7.13.5
Faker==40.15.0
h11==0.16.0
idna==3.13
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==26.2
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
PySocks==1.7.1
pytest==9.0.3
pytest-cov==7.1.0
requests==2.33.1
selenium==4.43.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
